Collapse MeasureSpec / ExploreSpec / OptimizeSpec into Experiment
The three experiment-flavoured spec classes are now a single
Experiment type with three static factories:

    Experiment.measuring(sampling, factors)
    Experiment.exploring(sampling)
    Experiment.optimizing(sampling)

The Spec sealed permits clause shrinks from four to two
(Experiment, ProbabilisticTest), mirroring the binary distinction
already present on the result side (ExperimentResult vs
ProbabilisticTestResult): experiments observe and produce
artefacts; tests assert and produce verdicts.

Author-facing effect:

    @PunitExperiment Experiment shoppingBaseline() { ... }
    @PunitExperiment Experiment shoppingComparison() { ... }
    @PunitExperiment Experiment temperatureSearch() { ... }

The kind lives on the verb in the body — measuring / exploring /
optimizing — where it reads more naturally than as a class name.
For tests / framework code that need to introspect, Experiment.kind()
returns a Kind enum (MEASURE / EXPLORE / OPTIMIZE).

Internals:
- Experiment holds a kind tag plus a typed Internal<?, ?, ?>
  delegate. Three static-nested Internal subclasses
  (MeasureInternal, ExploreInternal, OptimizeInternal) carry the
  kind-specific state and TypedSpec<FT, IT, OT> impl.
- Three Builder types (MeasureBuilder, ExploreBuilder,
  OptimizeBuilder) — typed at composition time, all returning a
  non-generic Experiment from .build().
- Engine dispatch is unchanged — Spec.dispatch routes through the
  Internal delegate's TypedSpec view.

Diagnostic accessors on Experiment populated by kind:
- lastSummary() — for measure runs; empty otherwise.
- history() — for optimize runs; empty otherwise.

Deleted:
- MeasureSpec.java
- ExploreSpec.java
- OptimizeSpec.java

Updated:
- Spec.java sealed permits = { Experiment, ProbabilisticTest }
- All call sites in tests and supporting code (Sampling javadoc,
  PowerAnalysis, BernoulliPassRate, PercentileLatency, etc.).
