+++ punit-api/src/main/java/org/javai/punit/api/typed/spec/Experiment.java
@@ -1,0 +1,550 @@
+package org.javai.punit.api.typed.spec;
+
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.time.Duration;
+import java.time.Instant;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Iterator;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.NoSuchElementException;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.OptionalLong;
+import java.util.function.Function;
+
+import org.javai.punit.api.typed.FactorBundle;
+import org.javai.punit.api.typed.Sampling;
+import org.javai.punit.api.typed.UseCase;
+import org.javai.punit.api.typed.ValueMatcher;
+import org.javai.punit.api.typed.spec.FactorMutator.IterationResult;
+
+/**
+ * The unified spec for the three experiment kinds — measure, explore,
+ * and optimize. The kind is chosen at the static factory:
+ *
+ * <ul>
+ *   <li>{@link #measuring(Sampling, Object)} — one configuration,
+ *       produces a baseline.</li>
+ *   <li>{@link #exploring(Sampling)} — many configurations, produces
+ *       a grid.</li>
+ *   <li>{@link #optimizing(Sampling)} — iterative search, produces an
+ *       optimization history.</li>
+ * </ul>
+ *
+ * <p>The engine treats all three uniformly via the
+ * {@link Spec#dispatch(Spec.Dispatcher)} protocol; the kind-specific
+ * configuration generation, summary aggregation, and result
+ * production live behind a typed internal delegate per kind.
+ *
+ * <p>The public class carries no type parameters — composition-time
+ * type safety lives on the typed builder, which the static factories
+ * return. The triplet of static factories on this single class
+ * replaces the historical
+ * {@code MeasureSpec} / {@code ExploreSpec} / {@code OptimizeSpec}
+ * classes; the experiment family is a single sealed permit on
+ * {@link Spec}, mirroring the binary
+ * {@code ExperimentResult} / {@code ProbabilisticTestResult} on the
+ * result side.
+ */
+public final class Experiment implements Spec {
+
+    /** Which experiment shape this value represents. */
+    public enum Kind { MEASURE, EXPLORE, OPTIMIZE }
+
+    private final Kind kind;
+    private final Internal<?, ?, ?> internal;
+
+    private Experiment(Kind kind, Internal<?, ?, ?> internal) {
+        this.kind = kind;
+        this.internal = internal;
+    }
+
+    // ── Static factories ─────────────────────────────────────────────
+
+    /**
+     * Compose a measure experiment over a factor-free
+     * {@link Sampling} and the factor bundle the measure runs against.
+     */
+    public static <FT, IT, OT> MeasureBuilder<FT, IT, OT> measuring(
+            Sampling<FT, IT, OT> sampling, FT factors) {
+        return new MeasureBuilder<>(
+                Objects.requireNonNull(sampling, "sampling"),
+                Objects.requireNonNull(factors, "factors"));
+    }
+
+    /**
+     * Compose an explore experiment over a {@link Sampling}; the
+     * factor grid is supplied through the returned builder.
+     */
+    public static <FT, IT, OT> ExploreBuilder<FT, IT, OT> exploring(
+            Sampling<FT, IT, OT> sampling) {
+        return new ExploreBuilder<>(Objects.requireNonNull(sampling, "sampling"));
+    }
+
+    /**
+     * Compose an optimize experiment over a {@link Sampling}; the
+     * initial factors, mutator, and scorer/direction are supplied
+     * through the returned builder.
+     */
+    public static <FT, IT, OT> OptimizeBuilder<FT, IT, OT> optimizing(
+            Sampling<FT, IT, OT> sampling) {
+        return new OptimizeBuilder<>(Objects.requireNonNull(sampling, "sampling"));
+    }
+
+    // ── Public scalar accessors ─────────────────────────────────────
+
+    public Kind kind() { return kind; }
+    public String experimentId() { return internal.experimentId(); }
+    public int samples() { return internal.samples(); }
+
+    /**
+     * Diagnostic accessor populated only after a measure experiment
+     * has run. Empty for explore and optimize, and for measures that
+     * have not yet been executed.
+     */
+    public Optional<SampleSummary<?>> lastSummary() {
+        if (internal instanceof MeasureInternal<?, ?, ?> m) {
+            return Optional.ofNullable(m.lastSummary.orElse(null));
+        }
+        return Optional.empty();
+    }
+
+    /**
+     * Diagnostic accessor populated only by optimize experiments.
+     * Empty for measure and explore.
+     */
+    public List<IterationResult<?>> history() {
+        if (internal instanceof OptimizeInternal<?, ?, ?> o) {
+            return List.copyOf(o.history);
+        }
+        return List.of();
+    }
+
+    // ── Spec dispatch ────────────────────────────────────────────────
+
+    @Override
+    public <R> R dispatch(Dispatcher<R> dispatcher) {
+        return doDispatch(internal, dispatcher);
+    }
+
+    private static <FT, IT, OT, R> R doDispatch(Internal<FT, IT, OT> typed, Dispatcher<R> d) {
+        return d.apply(typed);
+    }
+
+    // ── Common typed-internal base ───────────────────────────────────
+
+    private abstract static class Internal<FT, IT, OT> implements TypedSpec<FT, IT, OT> {
+
+        final Sampling<FT, IT, OT> sampling;
+        final String experimentId;
+
+        Internal(Sampling<FT, IT, OT> sampling, String experimentId) {
+            this.sampling = sampling;
+            this.experimentId = experimentId;
+        }
+
+        String experimentId() { return experimentId; }
+        int samples() { return sampling.samples(); }
+
+        @Override public Function<FT, UseCase<FT, IT, OT>> useCaseFactory() {
+            return sampling.useCaseFactory();
+        }
+
+        @Override public Optional<Duration> timeBudget() { return sampling.timeBudget(); }
+        @Override public OptionalLong tokenBudget() { return sampling.tokenBudget(); }
+        @Override public long tokenCharge() { return sampling.tokenCharge(); }
+        @Override public BudgetExhaustionPolicy budgetPolicy() { return sampling.budgetPolicy(); }
+        @Override public ExceptionPolicy exceptionPolicy() { return sampling.exceptionPolicy(); }
+        @Override public int maxExampleFailures() { return sampling.maxExampleFailures(); }
+    }
+
+    // ── Measure ──────────────────────────────────────────────────────
+
+    private static final class MeasureInternal<FT, IT, OT> extends Internal<FT, IT, OT> {
+
+        private final FT factors;
+        private final List<OT> expectedOutputs;
+        private final Optional<ValueMatcher<OT>> matcher;
+        private final Optional<Integer> expiresInDays;
+
+        private Optional<SampleSummary<OT>> lastSummary = Optional.empty();
+
+        private MeasureInternal(MeasureBuilder<FT, IT, OT> b) {
+            super(b.sampling, b.experimentId != null ? b.experimentId : defaultId("measure"));
+            this.factors = b.factors;
+            this.expectedOutputs = b.expected;
+            this.matcher = Optional.ofNullable(b.matcher);
+            this.expiresInDays = Optional.ofNullable(b.expiresInDays);
+        }
+
+        @Override public Optional<ValueMatcher<OT>> matcher() { return matcher; }
+
+        @Override public Iterator<Configuration<FT, IT, OT>> configurations() {
+            return List.of(new Configuration<>(
+                    factors, sampling.inputs(), expectedOutputs, sampling.samples())).iterator();
+        }
+
+        @Override public void consume(Configuration<FT, IT, OT> config, SampleSummary<OT> summary) {
+            this.lastSummary = Optional.of(summary);
+        }
+
+        @Override public EngineResult conclude() {
+            FactorBundle bundle = FactorBundle.of(factors);
+            Path path = defaultBaselinePath(experimentId, bundle);
+            String message = "measure baseline (stage 2 placeholder — serialisation lands in stage 4); "
+                    + "samples=" + lastSummary.map(SampleSummary::total).orElse(0)
+                    + ", passRate=" + lastSummary.map(s -> String.format("%.3f", s.passRate())).orElse("n/a");
+            return new ExperimentResult(message, path);
+        }
+    }
+
+    public static final class MeasureBuilder<FT, IT, OT> {
+
+        private final Sampling<FT, IT, OT> sampling;
+        private final FT factors;
+        private List<OT> expected = List.of();
+        private ValueMatcher<OT> matcher;
+        private String experimentId;
+        private Integer expiresInDays;
+
+        private MeasureBuilder(Sampling<FT, IT, OT> sampling, FT factors) {
+            this.sampling = sampling;
+            this.factors = factors;
+        }
+
+        public MeasureBuilder<FT, IT, OT> experimentId(String id) {
+            this.experimentId = Objects.requireNonNull(id, "experimentId");
+            return this;
+        }
+
+        /**
+         * Expected outputs, parallel to {@code sampling.inputs()}.
+         * Enables instance-conformance checking when supplied. A
+         * length mismatch is rejected at {@link #build()} with
+         * {@link IllegalStateException}.
+         */
+        public MeasureBuilder<FT, IT, OT> expectedOutputs(List<OT> outputs) {
+            Objects.requireNonNull(outputs, "outputs");
+            this.expected = List.copyOf(outputs);
+            return this;
+        }
+
+        @SafeVarargs
+        public final MeasureBuilder<FT, IT, OT> expectedOutputs(OT... outputs) {
+            return expectedOutputs(List.of(outputs));
+        }
+
+        /**
+         * Instance-conformance matcher, invoked per sample when
+         * {@code expectedOutputs} is supplied. Defaults to
+         * {@link ValueMatcher#equality()}.
+         */
+        public MeasureBuilder<FT, IT, OT> matcher(ValueMatcher<OT> matcher) {
+            this.matcher = Objects.requireNonNull(matcher, "matcher");
+            return this;
+        }
+
+        /** Baseline validity window. {@code 0} means no expiration. */
+        public MeasureBuilder<FT, IT, OT> expiresInDays(int days) {
+            if (days < 0) {
+                throw new IllegalArgumentException("expiresInDays must be non-negative, got " + days);
+            }
+            this.expiresInDays = days;
+            return this;
+        }
+
+        public Experiment build() {
+            if (!expected.isEmpty() && expected.size() != sampling.inputs().size()) {
+                throw new IllegalStateException(
+                        "expectedOutputs (" + expected.size() + ") and sampling.inputs() ("
+                                + sampling.inputs().size() + ") must be the same length");
+            }
+            if (!expected.isEmpty() && matcher == null) {
+                matcher = ValueMatcher.equality();
+            }
+            return new Experiment(Kind.MEASURE, new MeasureInternal<>(this));
+        }
+    }
+
+    // ── Explore ──────────────────────────────────────────────────────
+
+    private static final class ExploreInternal<FT, IT, OT> extends Internal<FT, IT, OT> {
+
+        private final List<FT> factors;
+        private final Map<FT, SampleSummary<OT>> perConfig = new LinkedHashMap<>();
+
+        private ExploreInternal(ExploreBuilder<FT, IT, OT> b) {
+            super(b.sampling, b.experimentId != null ? b.experimentId : defaultId("explore"));
+            this.factors = b.factors;
+        }
+
+        @Override public Iterator<Configuration<FT, IT, OT>> configurations() {
+            List<Configuration<FT, IT, OT>> configs = new ArrayList<>(factors.size());
+            for (FT ft : factors) {
+                configs.add(Configuration.of(ft, sampling.inputs(), sampling.samples()));
+            }
+            return configs.iterator();
+        }
+
+        @Override public void consume(Configuration<FT, IT, OT> config, SampleSummary<OT> summary) {
+            perConfig.put(config.factors(), summary);
+        }
+
+        @Override public EngineResult conclude() {
+            Path dir = Paths.get("explorations", experimentId);
+            String message = "explore artefact (stage 2 placeholder); configurations="
+                    + perConfig.size();
+            return new ExperimentResult(message, dir);
+        }
+    }
+
+    public static final class ExploreBuilder<FT, IT, OT> {
+
+        private final Sampling<FT, IT, OT> sampling;
+        private List<FT> factors;
+        private String experimentId;
+
+        private ExploreBuilder(Sampling<FT, IT, OT> sampling) {
+            this.sampling = sampling;
+        }
+
+        public ExploreBuilder<FT, IT, OT> factors(List<FT> grid) {
+            Objects.requireNonNull(grid, "factors");
+            if (grid.isEmpty()) {
+                throw new IllegalArgumentException("factors must be non-empty");
+            }
+            this.factors = List.copyOf(grid);
+            return this;
+        }
+
+        @SafeVarargs
+        public final ExploreBuilder<FT, IT, OT> factors(FT... grid) {
+            Objects.requireNonNull(grid, "factors");
+            return factors(List.of(grid));
+        }
+
+        public ExploreBuilder<FT, IT, OT> experimentId(String id) {
+            this.experimentId = Objects.requireNonNull(id, "experimentId");
+            return this;
+        }
+
+        public Experiment build() {
+            if (factors == null) {
+                throw new IllegalStateException("factors is required — call .factors(...)");
+            }
+            return new Experiment(Kind.EXPLORE, new ExploreInternal<>(this));
+        }
+    }
+
+    // ── Optimize ─────────────────────────────────────────────────────
+
+    private static final class OptimizeInternal<FT, IT, OT> extends Internal<FT, IT, OT> {
+
+        private final FT initialFactors;
+        private final FactorMutator<FT> mutator;
+        private final Scorer scorer;
+        private final boolean maximizing;
+        private final int maxIterations;
+        private final int noImprovementWindow;
+
+        private final List<IterationResult<FT>> history = new ArrayList<>();
+
+        private OptimizeInternal(OptimizeBuilder<FT, IT, OT> b) {
+            super(b.sampling, b.experimentId != null ? b.experimentId : defaultId("optimize"));
+            this.initialFactors = b.initialFactors;
+            this.mutator = b.mutator;
+            this.scorer = b.scorer;
+            this.maximizing = b.maximizing;
+            this.maxIterations = b.maxIterations;
+            this.noImprovementWindow = b.noImprovementWindow;
+        }
+
+        @Override public Iterator<Configuration<FT, IT, OT>> configurations() {
+            return new AdaptiveIterator();
+        }
+
+        @Override public void consume(Configuration<FT, IT, OT> config, SampleSummary<OT> summary) {
+            double score = scorer.score(summary);
+            history.add(new IterationResult<>(config.factors(), score));
+        }
+
+        @Override public EngineResult conclude() {
+            IterationResult<FT> best = bestSoFar();
+            Path dir = Paths.get("optimizations", experimentId);
+            String message = "optimize artefact (stage 2 placeholder); iterations="
+                    + history.size()
+                    + (best == null ? "" : ", bestScore=" + String.format("%.3f", best.score()));
+            return new ExperimentResult(message, dir);
+        }
+
+        private IterationResult<FT> bestSoFar() {
+            IterationResult<FT> best = null;
+            for (IterationResult<FT> r : history) {
+                if (best == null || isBetter(r.score(), best.score())) {
+                    best = r;
+                }
+            }
+            return best;
+        }
+
+        private boolean isBetter(double a, double b) {
+            return maximizing ? a > b : a < b;
+        }
+
+        private final class AdaptiveIterator implements Iterator<Configuration<FT, IT, OT>> {
+
+            private FT next = initialFactors;
+            private int issued = 0;
+            private int iterationsSinceImprovement = 0;
+            private double bestScoreSeen = Double.NaN;
+
+            @Override public boolean hasNext() {
+                refresh();
+                return next != null;
+            }
+
+            @Override public Configuration<FT, IT, OT> next() {
+                refresh();
+                if (next == null) {
+                    throw new NoSuchElementException();
+                }
+                FT factors = next;
+                next = null;
+                issued++;
+                return Configuration.of(factors, sampling.inputs(), sampling.samples());
+            }
+
+            private void refresh() {
+                if (next != null) {
+                    return;
+                }
+                if (issued >= maxIterations) {
+                    return;
+                }
+                if (!history.isEmpty()) {
+                    IterationResult<FT> last = history.get(history.size() - 1);
+                    if (Double.isNaN(bestScoreSeen)
+                            || isBetter(last.score(), bestScoreSeen)) {
+                        bestScoreSeen = last.score();
+                        iterationsSinceImprovement = 0;
+                    } else {
+                        iterationsSinceImprovement++;
+                    }
+                    if (iterationsSinceImprovement >= noImprovementWindow) {
+                        return;
+                    }
+                    FT current = last.factors();
+                    FT candidate = mutator.next(current,
+                            Collections.unmodifiableList(history));
+                    next = candidate;
+                }
+            }
+        }
+    }
+
+    public static final class OptimizeBuilder<FT, IT, OT> {
+
+        private final Sampling<FT, IT, OT> sampling;
+        private FT initialFactors;
+        private FactorMutator<FT> mutator;
+        private Scorer scorer;
+        private boolean maximizing;
+        private boolean directionSet;
+        private int maxIterations = 20;
+        private int noImprovementWindow = 5;
+        private String experimentId;
+
+        private OptimizeBuilder(Sampling<FT, IT, OT> sampling) {
+            this.sampling = sampling;
+        }
+
+        public OptimizeBuilder<FT, IT, OT> initialFactors(FT factors) {
+            this.initialFactors = Objects.requireNonNull(factors, "initialFactors");
+            return this;
+        }
+
+        public OptimizeBuilder<FT, IT, OT> mutator(FactorMutator<FT> m) {
+            this.mutator = Objects.requireNonNull(m, "mutator");
+            return this;
+        }
+
+        /**
+         * Sets the scorer and declares the optimisation direction as
+         * "maximise the score." Mutually exclusive with
+         * {@link #minimize(Scorer)} — last one called wins. One of the
+         * two is required.
+         */
+        public OptimizeBuilder<FT, IT, OT> maximize(Scorer scorer) {
+            this.scorer = Objects.requireNonNull(scorer, "scorer");
+            this.maximizing = true;
+            this.directionSet = true;
+            return this;
+        }
+
+        /**
+         * Sets the scorer and declares the optimisation direction as
+         * "minimise the score." Mutually exclusive with
+         * {@link #maximize(Scorer)} — last one called wins. One of the
+         * two is required.
+         */
+        public OptimizeBuilder<FT, IT, OT> minimize(Scorer scorer) {
+            this.scorer = Objects.requireNonNull(scorer, "scorer");
+            this.maximizing = false;
+            this.directionSet = true;
+            return this;
+        }
+
+        public OptimizeBuilder<FT, IT, OT> maxIterations(int n) {
+            if (n < 1) {
+                throw new IllegalArgumentException("maxIterations must be >= 1");
+            }
+            this.maxIterations = n;
+            return this;
+        }
+
+        public OptimizeBuilder<FT, IT, OT> noImprovementWindow(int n) {
+            if (n < 1) {
+                throw new IllegalArgumentException("noImprovementWindow must be >= 1");
+            }
+            this.noImprovementWindow = n;
+            return this;
+        }
+
+        public OptimizeBuilder<FT, IT, OT> experimentId(String id) {
+            this.experimentId = Objects.requireNonNull(id, "experimentId");
+            return this;
+        }
+
+        public Experiment build() {
+            if (initialFactors == null) {
+                throw new IllegalStateException("initialFactors is required");
+            }
+            if (mutator == null) {
+                throw new IllegalStateException("mutator is required");
+            }
+            if (!directionSet) {
+                throw new IllegalStateException(
+                        "scorer is required — call .maximize(...) or .minimize(...)");
+            }
+            return new Experiment(Kind.OPTIMIZE, new OptimizeInternal<>(this));
+        }
+    }
+
+    // ── Helpers ──────────────────────────────────────────────────────
+
+    private static String defaultId(String kind) {
+        return kind + "-" + Instant.now().toEpochMilli();
+    }
+
+    static Path defaultBaselinePath(String experimentId, FactorBundle bundle) {
+        String filename = experimentId;
+        if (!bundle.isEmpty()) {
+            filename = filename + "-" + bundle.bundleHash();
+        }
+        return Paths.get("specs", filename + ".yaml");
+    }
+}
