+++ punit-api/src/main/java/org/javai/punit/api/typed/spec/Scorer.java
@@ -2,7 +2,7 @@ package org.javai.punit.api.typed.spec;
 
 /**
  * Reduces a {@link SampleSummary} to a single comparable score for an
- * {@link OptimizeSpec} iteration.
+ * {@link Experiment} iteration.
  */
 @FunctionalInterface
 public interface Scorer {
