+++ punit-api/src/test/java/org/javai/punit/api/typed/architecture/PunitApiArchitectureTest.java
@@ -97,11 +97,11 @@ class PunitApiArchitectureTest {
     void specsExposeOnlyVerbFormFactories() {
         ArchRule rule = noMethods()
                 .that().areDeclaredInClassesThat()
-                .haveFullyQualifiedName("org.javai.punit.api.typed.spec.MeasureSpec")
+                .haveFullyQualifiedName("org.javai.punit.api.typed.spec.Experiment")
                 .or().areDeclaredInClassesThat()
-                .haveFullyQualifiedName("org.javai.punit.api.typed.spec.ExploreSpec")
+                .haveFullyQualifiedName("org.javai.punit.api.typed.spec.Experiment")
                 .or().areDeclaredInClassesThat()
-                .haveFullyQualifiedName("org.javai.punit.api.typed.spec.OptimizeSpec")
+                .haveFullyQualifiedName("org.javai.punit.api.typed.spec.Experiment")
                 .or().areDeclaredInClassesThat()
                 .haveFullyQualifiedName("org.javai.punit.api.typed.spec.ProbabilisticTest")
                 .should().haveName("builder")
@@ -114,9 +114,9 @@ class PunitApiArchitectureTest {
     @Test
     @DisplayName("each spec exposes its verb-form factory")
     void specsHaveVerbFormFactory() {
-        verbFormFactory("MeasureSpec", "measuring");
-        verbFormFactory("ExploreSpec", "exploring");
-        verbFormFactory("OptimizeSpec", "optimizing");
+        verbFormFactory("Experiment", "measuring");
+        verbFormFactory("Experiment", "exploring");
+        verbFormFactory("Experiment", "optimizing");
         verbFormFactory("ProbabilisticTest", "testing");
     }
